refactor(erd): extract pure utilities and constants from ERDView
